backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "openai==1.3.6",
    "cohere==5.5.3",
    "qdrant-client==1.8.0",
    "python-dotenv==1.0.0",
    "numpy==1.26.0",
    "pydantic==2.7.0",
    "pydantic-settings==2.12.0",
    "requests==2.31.0",
    "slowapi==0.1.9",
    "psycopg2-binary==2.9.9",
    "pgvector==1.3.0"
]

[tool.black]
line-length = 88
target-version = ['py314']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''